refactored etools-logo and code adjustments
