+++ polymer_3/src_ts/elements/etools-logo.ts
@@ -1,22 +1,21 @@
 import {html} from '@polymer/polymer';
 
+
 export const etoolsLogo = html`
-  <style>
-    :host {
-      line-height: 1;
-    }
-  </style>
-
-  <svg xmlns="http://www.w3.org/2000/svg" width$="[[size]]" viewBox="0 0 363.7 105.4" enable-background="new 0 0 363.7 105.4"><path fill="#00A551" d="M50.9 96v-45.2h45.2z"/><path fill="#1C85BF" d="M50.9 5.6v45.2h-45.2z"/><path fill="#00ADEE" d="M50.9 50.8v-45.2h45.2z"/><path fill="#8BC53F" d="M96.1 5.6l-.1 45.2h-45.1z"/><g fill$="[[textColor]]"><path id="svg_12" class="st5" d="M122 56c0 5.4 1.2 9.4 3.7 11.8 2.5 2.5 6.3 3.7 11.3 3.7 2.5 0 4.6-.5 6.3-1.5s3.3-2.4 4.6-4.2l7 2.9c-1.8 3-4.3 5.4-7.5 7.1s-6.8 2.6-10.9 2.6-7.6-.6-10.5-1.7c-2.9-1.1-5.2-2.7-7-4.8-1.8-2.1-3.1-4.6-3.9-7.5s-1.2-6.2-1.2-9.8v-1c0-1.5 0-3 .1-4.5.1-1.4.2-2.9.6-4.3.3-1.4.8-2.8 1.5-4.2s1.6-2.8 2.9-4.1c2.1-2.3 4.6-4.1 7.5-5.2 2.9-1.1 5.9-1.7 9-1.7 3.5 0 6.5.6 9.2 1.7s4.9 2.7 6.7 4.6c1.8 2 3.2 4.3 4.1 7 .9 2.7 1.4 5.7 1.4 9v4.2h-34.9v-.1zm0-6.5h27v-.8c0-1.5-.3-3.1-.9-4.5-.6-1.5-1.5-2.8-2.7-3.9-1.2-1.1-2.6-2.1-4.3-2.8s-3.5-1.1-5.6-1.1c-2 0-3.8.3-5.5 1s-3.1 1.6-4.3 2.7c-1.2 1.1-2.1 2.4-2.8 3.9s-1 3-1 4.6v.9h.1z"/><path id="svg_13" class="st5" d="M182.8 20.6v56.7h-8.3v-56.7h-21.4v-7.6h51.1l-2.4 7.6h-19z"/><path id="svg_14" class="st5" d="M236.3 71.7c-2.2 2.3-4.6 4-7.3 5-2.7 1.1-5.6 1.6-8.8 1.6-3 0-5.9-.6-8.7-1.7s-5.2-2.8-7.4-5c-2.3-2.3-3.9-5.1-4.6-8.1-.7-3.1-1.1-6.3-1.1-9.6v-1.8c.1-3.2.5-6.1 1.2-8.6s2.3-5.1 4.7-7.6c2.2-2.2 4.6-3.9 7.2-4.9 2.6-1 5.5-1.6 8.6-1.6s5.9.5 8.6 1.5c2.6 1 5 2.6 7.2 4.8 2.2 2.3 3.8 4.7 4.8 7.4 1 2.6 1.5 5.6 1.5 8.8v1.9c0 3.3-.4 6.5-1.1 9.4-.8 3.1-2.4 5.9-4.8 8.5zm-1.5-17.8v-1.8c0-4.5-1.2-8.2-3.5-11-1.2-1.5-2.8-2.6-4.6-3.5s-4-1.2-6.4-1.2c-2.3 0-4.4.4-6.2 1.2s-3.3 1.9-4.5 3.3c-2.5 2.8-3.7 6.6-3.7 11.4v2.1c0 2.6.3 5 .8 7.2.6 2.2 1.5 4.1 2.9 5.7 1.1 1.4 2.5 2.4 4.3 3.1s3.9 1.1 6.4 1.1c2.3 0 4.3-.3 6-1s3.1-1.6 4.3-2.9c1.6-1.6 2.7-3.6 3.3-6s.9-4.9.9-7.7z"/><path id="svg_15" class="st5" d="M288.8 71.7c-2.2 2.3-4.6 4-7.3 5-2.7 1.1-5.6 1.6-8.8 1.6-3 0-5.9-.6-8.7-1.7s-5.2-2.8-7.4-5c-2.3-2.3-3.9-5.1-4.6-8.1-.7-3.1-1.1-6.3-1.1-9.6v-1.8c.1-3.2.5-6.1 1.2-8.6s2.3-5.1 4.7-7.6c2.2-2.2 4.6-3.9 7.2-4.9s5.5-1.6 8.6-1.6 5.9.5 8.6 1.5c2.6 1 5 2.6 7.2 4.8 2.2 2.3 3.8 4.7 4.8 7.4 1 2.6 1.5 5.6 1.5 8.8v1.9c0 3.3-.4 6.5-1.1 9.4-.8 3.1-2.4 5.9-4.8 8.5zm-1.5-17.8v-1.8c0-4.5-1.2-8.2-3.5-11-1.2-1.5-2.8-2.6-4.6-3.5s-4-1.2-6.4-1.2c-2.3 0-4.4.4-6.2 1.2-1.8.8-3.3 1.9-4.5 3.3-2.5 2.8-3.7 6.6-3.7 11.4v2.1c0 2.6.3 5 .8 7.2.6 2.2 1.5 4.1 2.9 5.7 1.1 1.4 2.5 2.4 4.3 3.1s3.9 1.1 6.4 1.1c2.3 0 4.3-.3 6-1s3.1-1.6 4.3-2.9c1.6-1.6 2.7-3.6 3.3-6 .6-2.4.9-4.9.9-7.7z"/><path id="svg_16" class="st5" d="M304 14.1l7.9-2.2v54.1c0 3.7 1.9 5.5 5.7 5.5h4.7v5.7c-1 .4-2.2.6-3.4.8-1.2.2-2.2.3-3.1.3-1.7 0-3.2-.3-4.7-1-1.4-.6-2.7-1.5-3.8-2.6-1.1-1.1-1.9-2.4-2.5-3.9s-.9-3.2-.9-4.9l.1-51.8z"/><path id="svg_17" class="st5" d="M345.8 29.5c3 0 5.8.6 8.6 1.8 2.8 1.2 5.1 2.8 7.1 5l-4.8 4.7c-1-1.1-2.4-2.2-4.2-3.1-1.7-1-4-1.5-6.9-1.5-1 0-2 .1-3.2.2-1.1.2-2.2.5-3.2.9-1 .5-1.9 1.1-2.5 2-.7.9-1 2-1 3.3 0 1.5.7 2.7 2.1 3.6s3.1 1.6 5.2 2.2c2.1.6 4.3 1.3 6.7 1.8 2.4.6 4.7 1.4 6.7 2.3 2.1 1 3.8 2.2 5.2 3.8 1.4 1.6 2.1 3.7 2.1 6.2 0 4.8-1.6 8.6-4.7 11.4s-7.8 4.2-13.9 4.2c-4.6 0-8.5-.8-11.6-2.4s-5.7-4.1-7.6-7.4l5.7-3.6c.8.9 1.6 1.8 2.5 2.6.9.8 1.8 1.5 2.9 2.1 1 .6 2.3 1 3.7 1.4 1.4.3 3 .5 4.9.5 1 0 2.1-.1 3.3-.3 1.2-.2 2.3-.6 3.3-1.1 1-.6 1.9-1.3 2.5-2.3.7-1 1-2.2 1-3.8 0-1.8-.7-3.2-2.1-4.2-1.4-.9-3.1-1.7-5.2-2.3-2.1-.6-4.3-1.1-6.7-1.6-2.4-.5-4.6-1.2-6.7-2.2s-3.8-2.4-5.2-4.1c-1.4-1.8-2.1-4.2-2.1-7.3 0-2.4.6-4.4 1.8-6.1 1.2-1.6 2.6-3 4.4-4s3.7-1.7 5.8-2.2c2.2-.2 4.2-.5 6.1-.5z"/></g></svg>
-
+  <iron-iconset-svg name="etools-logo">
+    <svg xmlns="http://www.w3.org/2000/svg" width="36" viewBox="0 0 363.7 105.4" enable-background="new 0 0 363.7 105.4">
+        <path fill="#00A551" d="M50.9 96v-45.2h45.2z"/>
+        <path fill="#1C85BF" d="M50.9 5.6v45.2h-45.2z"/>
+        <path fill="#00ADEE" d="M50.9 50.8v-45.2h45.2z"/>
+        <path fill="#8BC53F" d="M96.1 5.6l-.1 45.2h-45.1z"/>
+        <g fill="#ffffff">
+          <path id="svg_12" class="st5" d="M122 56c0 5.4 1.2 9.4 3.7 11.8 2.5 2.5 6.3 3.7 11.3 3.7 2.5 0 4.6-.5 6.3-1.5s3.3-2.4 4.6-4.2l7 2.9c-1.8 3-4.3 5.4-7.5 7.1s-6.8 2.6-10.9 2.6-7.6-.6-10.5-1.7c-2.9-1.1-5.2-2.7-7-4.8-1.8-2.1-3.1-4.6-3.9-7.5s-1.2-6.2-1.2-9.8v-1c0-1.5 0-3 .1-4.5.1-1.4.2-2.9.6-4.3.3-1.4.8-2.8 1.5-4.2s1.6-2.8 2.9-4.1c2.1-2.3 4.6-4.1 7.5-5.2 2.9-1.1 5.9-1.7 9-1.7 3.5 0 6.5.6 9.2 1.7s4.9 2.7 6.7 4.6c1.8 2 3.2 4.3 4.1 7 .9 2.7 1.4 5.7 1.4 9v4.2h-34.9v-.1zm0-6.5h27v-.8c0-1.5-.3-3.1-.9-4.5-.6-1.5-1.5-2.8-2.7-3.9-1.2-1.1-2.6-2.1-4.3-2.8s-3.5-1.1-5.6-1.1c-2 0-3.8.3-5.5 1s-3.1 1.6-4.3 2.7c-1.2 1.1-2.1 2.4-2.8 3.9s-1 3-1 4.6v.9h.1z"/>
+          <path id="svg_13" class="st5" d="M182.8 20.6v56.7h-8.3v-56.7h-21.4v-7.6h51.1l-2.4 7.6h-19z"/>
+          <path id="svg_14" class="st5" d="M236.3 71.7c-2.2 2.3-4.6 4-7.3 5-2.7 1.1-5.6 1.6-8.8 1.6-3 0-5.9-.6-8.7-1.7s-5.2-2.8-7.4-5c-2.3-2.3-3.9-5.1-4.6-8.1-.7-3.1-1.1-6.3-1.1-9.6v-1.8c.1-3.2.5-6.1 1.2-8.6s2.3-5.1 4.7-7.6c2.2-2.2 4.6-3.9 7.2-4.9 2.6-1 5.5-1.6 8.6-1.6s5.9.5 8.6 1.5c2.6 1 5 2.6 7.2 4.8 2.2 2.3 3.8 4.7 4.8 7.4 1 2.6 1.5 5.6 1.5 8.8v1.9c0 3.3-.4 6.5-1.1 9.4-.8 3.1-2.4 5.9-4.8 8.5zm-1.5-17.8v-1.8c0-4.5-1.2-8.2-3.5-11-1.2-1.5-2.8-2.6-4.6-3.5s-4-1.2-6.4-1.2c-2.3 0-4.4.4-6.2 1.2s-3.3 1.9-4.5 3.3c-2.5 2.8-3.7 6.6-3.7 11.4v2.1c0 2.6.3 5 .8 7.2.6 2.2 1.5 4.1 2.9 5.7 1.1 1.4 2.5 2.4 4.3 3.1s3.9 1.1 6.4 1.1c2.3 0 4.3-.3 6-1s3.1-1.6 4.3-2.9c1.6-1.6 2.7-3.6 3.3-6s.9-4.9.9-7.7z"/>
+          <path id="svg_15" class="st5" d="M288.8 71.7c-2.2 2.3-4.6 4-7.3 5-2.7 1.1-5.6 1.6-8.8 1.6-3 0-5.9-.6-8.7-1.7s-5.2-2.8-7.4-5c-2.3-2.3-3.9-5.1-4.6-8.1-.7-3.1-1.1-6.3-1.1-9.6v-1.8c.1-3.2.5-6.1 1.2-8.6s2.3-5.1 4.7-7.6c2.2-2.2 4.6-3.9 7.2-4.9s5.5-1.6 8.6-1.6 5.9.5 8.6 1.5c2.6 1 5 2.6 7.2 4.8 2.2 2.3 3.8 4.7 4.8 7.4 1 2.6 1.5 5.6 1.5 8.8v1.9c0 3.3-.4 6.5-1.1 9.4-.8 3.1-2.4 5.9-4.8 8.5zm-1.5-17.8v-1.8c0-4.5-1.2-8.2-3.5-11-1.2-1.5-2.8-2.6-4.6-3.5s-4-1.2-6.4-1.2c-2.3 0-4.4.4-6.2 1.2-1.8.8-3.3 1.9-4.5 3.3-2.5 2.8-3.7 6.6-3.7 11.4v2.1c0 2.6.3 5 .8 7.2.6 2.2 1.5 4.1 2.9 5.7 1.1 1.4 2.5 2.4 4.3 3.1s3.9 1.1 6.4 1.1c2.3 0 4.3-.3 6-1s3.1-1.6 4.3-2.9c1.6-1.6 2.7-3.6 3.3-6 .6-2.4.9-4.9.9-7.7z"/>
+          <path id="svg_16" class="st5" d="M304 14.1l7.9-2.2v54.1c0 3.7 1.9 5.5 5.7 5.5h4.7v5.7c-1 .4-2.2.6-3.4.8-1.2.2-2.2.3-3.1.3-1.7 0-3.2-.3-4.7-1-1.4-.6-2.7-1.5-3.8-2.6-1.1-1.1-1.9-2.4-2.5-3.9s-.9-3.2-.9-4.9l.1-51.8z"/>
+          <path id="svg_17" class="st5" d="M345.8 29.5c3 0 5.8.6 8.6 1.8 2.8 1.2 5.1 2.8 7.1 5l-4.8 4.7c-1-1.1-2.4-2.2-4.2-3.1-1.7-1-4-1.5-6.9-1.5-1 0-2 .1-3.2.2-1.1.2-2.2.5-3.2.9-1 .5-1.9 1.1-2.5 2-.7.9-1 2-1 3.3 0 1.5.7 2.7 2.1 3.6s3.1 1.6 5.2 2.2c2.1.6 4.3 1.3 6.7 1.8 2.4.6 4.7 1.4 6.7 2.3 2.1 1 3.8 2.2 5.2 3.8 1.4 1.6 2.1 3.7 2.1 6.2 0 4.8-1.6 8.6-4.7 11.4s-7.8 4.2-13.9 4.2c-4.6 0-8.5-.8-11.6-2.4s-5.7-4.1-7.6-7.4l5.7-3.6c.8.9 1.6 1.8 2.5 2.6.9.8 1.8 1.5 2.9 2.1 1 .6 2.3 1 3.7 1.4 1.4.3 3 .5 4.9.5 1 0 2.1-.1 3.3-.3 1.2-.2 2.3-.6 3.3-1.1 1-.6 1.9-1.3 2.5-2.3.7-1 1-2.2 1-3.8 0-1.8-.7-3.2-2.1-4.2-1.4-.9-3.1-1.7-5.2-2.3-2.1-.6-4.3-1.1-6.7-1.6-2.4-.5-4.6-1.2-6.7-2.2s-3.8-2.4-5.2-4.1c-1.4-1.8-2.1-4.2-2.1-7.3 0-2.4.6-4.4 1.8-6.1 1.2-1.6 2.6-3 4.4-4s3.7-1.7 5.8-2.2c2.2-.2 4.2-.5 6.1-.5z"/>
+        </g>
+    </svg>
+  </iron-iconset-svg>
 `;
-
-
-
-
-
-
-
-
-
-
