+++ polymer_3/src_ts/elements/etools-prp-reset.ts
@@ -2,7 +2,7 @@ import {PolymerElement} from '@polymer/polymer';
 import UtilsMixin from '../mixins/utils-mixin';
 import {Debouncer} from '@polymer/polymer/lib/utils/debounce';
 import {property} from '@polymer/decorators/lib/decorators';
-import {GenericObject} from "../typings/globals.types";
+import {GenericObject} from '../typings/globals.types';
 
 
 /**
